% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_json_files.R
\name{transform_json_files}
\alias{transform_json_files}
\title{Function that extract, transform and load data from Statistics Poland API}
\usage{
transform_json_files(number_of_cores)
}
\arguments{
\item{number_of_clusters}{number of CPU cores taking part in processing}
}
\description{
Function that extract, transform and load data from Statistics Poland API
}
